requirements.txt
aiohttp==3.8.4
aiohttp-retry==2.8.3
aiosignal==1.3.1
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
billiard==4.1.0
celery==5.3.1
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.5
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
dj-database-url==2.0.0
Django==4.2.3
django-cors-headers==4.2.0
django-filter==23.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.3
Faker==19.1.0
flower==2.0.0
frozenlist==1.4.0
humanize==4.7.0
idna==3.4
inflection==0.5.1
jsonschema==4.18.3
jsonschema-specifications==2023.6.1
kombu==5.3.1
Markdown==3.4.3
multidict==6.0.4
Pillow==10.0.0
prometheus-client==0.17.1
prompt-toolkit==3.0.39
psycopg2-binary==2.9.6
PyJWT==2.7.0
pyotp==2.8.0
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3
PyYAML==6.0
referencing==0.29.1
requests==2.31.0
rpds-py==0.8.10
six==1.16.0
sqlparse==0.4.4
tornado==6.3.2
twilio==8.5.0
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.3
vine==5.0.0
wcwidth==0.2.6
yarl==1.9.2
